+++ src/main/java/net/fabricmc/mappingio/MappingReader.java
@@ -33,6 +33,7 @@ import net.fabricmc.mappingio.format.enigma.EnigmaDirReader;
 import net.fabricmc.mappingio.format.enigma.EnigmaFileReader;
 import net.fabricmc.mappingio.format.proguard.ProGuardFileReader;
 import net.fabricmc.mappingio.format.simple.RecafSimpleFileReader;
+import net.fabricmc.mappingio.format.srg.JamFileReader;
 import net.fabricmc.mappingio.format.srg.SrgFileReader;
 import net.fabricmc.mappingio.format.srg.TsrgFileReader;
 import net.fabricmc.mappingio.format.tiny.Tiny1FileReader;
@@ -83,9 +84,14 @@ public final class MappingReader {
 			return MappingFormat.ENIGMA_FILE;
 		case "PK:":
 		case "CL:":
-		case "MD:":
 		case "FD:":
+		case "MD:":
 			return detectSrgOrXsrg(br);
+		case "CL ":
+		case "FD ":
+		case "MD ":
+		case "MP ":
+			return MappingFormat.JAM_FILE;
 		}
 
 		String headerStr = String.valueOf(buffer, 0, pos);
@@ -261,6 +267,9 @@ public final class MappingReader {
 		case SRG_FILE:
 		case XSRG_FILE:
 			SrgFileReader.read(reader, visitor);
+			break;
+		case JAM_FILE:
+			JamFileReader.read(reader, visitor);
 			break;
 		case CSRG_FILE:
 		case TSRG_FILE: