+++ src/test/java/net/fabricmc/mappingio/read/EmptyContentReadTest.java
@@ -26,6 +26,7 @@ import org.junit.jupiter.api.Test;
 import net.fabricmc.mappingio.format.enigma.EnigmaFileReader;
 import net.fabricmc.mappingio.format.proguard.ProGuardFileReader;
 import net.fabricmc.mappingio.format.simple.RecafSimpleFileReader;
+import net.fabricmc.mappingio.format.srg.JamFileReader;
 import net.fabricmc.mappingio.format.srg.SrgFileReader;
 import net.fabricmc.mappingio.format.srg.TsrgFileReader;
 import net.fabricmc.mappingio.format.tiny.Tiny1FileReader;
@@ -59,6 +60,11 @@ public class EmptyContentReadTest {
 	@Test
 	public void emptySrgFile() throws Exception {
 		SrgFileReader.read(new StringReader(""), tree);
+	}
+
+	@Test
+	public void emptyJamFile() throws Exception {
+		JamFileReader.read(new StringReader(""), tree);
 	}
 
 	@Test