+++ src/test/java/net/fabricmc/mappingio/visiting/VisitEndTest.java
@@ -77,6 +77,12 @@ public class VisitEndTest {
 	}
 
 	@Test
+	public void jamFile() throws Exception {
+		MappingFormat format = MappingFormat.JAM_FILE;
+		check(format);
+	}
+
+	@Test
 	public void csrgFile() throws Exception {
 		MappingFormat format = MappingFormat.CSRG_FILE;
 		check(format);